requirements.txt
asttokens==2.4.1
attrs==23.2.0
cattrs==23.2.3
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==5.1.1
executing==2.0.1
flatbuffers==24.3.25
idna==3.7
ipython==8.23.0
jedi==0.19.1
matplotlib-inline==0.1.6
numpy==1.26.4
openmeteo_requests==1.2.0
openmeteo_sdk==1.11.4
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
requests-cache==1.2.0
retry-requests==2.0.0
six==1.16.0
stack-data==0.6.3
traitlets==5.14.2
typing_extensions==4.11.0
tzdata==2024.1
url-normalize==1.4.3
urllib3==2.2.1
wcwidth==0.2.13
